Redo the tab selector UI to keep the glass morphism but have a cleaner and more minimalist and sleeker look.
Make the info description be readable. Currently it's hard to read some of it because of it overlaying on the tab selector text. I trust you to find a workaround. if you are unable to think of anything. Just don't make it a transparent background at all

Remove the blue outline for the selected text box. I don't want any indicator for when I select a text box AT ALL