the black outline flash is still there
Remove the padding for the area between the start of the app and the name of it. (where the descriptions were before we moved them into the icon)

Improve the design of the output result UI to be more clean and sleek. The text design and spacing looks bad right now